Break the `variance` module into submodules for ease of comprehension.
